pyproject.toml
[tool.poetry]
name = "kaizen-blog"
version = "0.1.0"
description = "The Kaizen blog"
authors = ["Alfonso <amlluch@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
boto3 = "^1.17.39"
kink = "^0.5.0"
aws-lambda-context = "^1.1.0"
marshmallow = "^3.10.0"
marshmallow-dataclass = "^8.3.2"
Pillow = "^8.1.2"
docker = "^4.4.4"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pre-commit = "^2.11.1"
isort = "^5.8.0"
flake8 = "^3.9.0"
pytest = "^6.2.2"
moto = "^2.0.2"
mypy = "^0.812"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ["py38"]
